packages/cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tracegrade"
version = "0.1.0"
description = "CLI for TraceGrade - turn AI agent failures into regression tests"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.scripts]
tracegrade = "tracegrade.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tracegrade"]
